+++ components/integrations/confirm_integration/index.js
@@ -6,7 +6,7 @@ import {connect} from 'react-redux';
 import {getCommands, getOAuthApps, getIncomingHooks, getOutgoingHooks} from 'mattermost-redux/selectors/entities/integrations';
 import {getBotAccounts} from 'mattermost-redux/selectors/entities/bots';
 
-import ConfirmIntegration from './confirm_integration.jsx';
+import ConfirmIntegration from './confirm_integration';
 
 function mapStateToProps(state) {
     return {
